+++ Sources/MarkersExtractor/Extractors/FCPXMLMarkerExtractor.swift
@@ -87,9 +87,64 @@ class FCPXMLMarkerExtractor: NSObject, ProgressReporting {
     }
     
     // MARK: - Public Instance Methods
-
+    
+    public struct TimelineContext {
+        public let library: FinalCutPro.FCPXML.Library?
+        public let projectName: String?
+        public let timeline: FinalCutPro.FCPXML.AnyTimeline
+        public let timelineName: String
+        public let timelineStartTime: Timecode
+    }
+    
+    /// Returns the first timeline found in the FCPXML as well as contextual metadata.
+    public func extractTimelineContext(
+        defaultTimelineName: String
+    ) -> TimelineContext? {
+        let parsedFCPXML = FinalCutPro.FCPXML(fileContent: fcpxmlDoc)
+        
+        let library = parsedFCPXML.root.library
+        
+        // prioritize a project if one exists, otherwise use clips
+        guard let timeline = parsedFCPXML.allTimelines().first else {
+            logger.info(
+                "No timelines (projects or clips) could be found in the FCPXML."
+            )
+            return nil
+        }
+        
+        // project element may or may not exist.
+        let parentProject = timeline.element
+            .ancestorElements(includingSelf: false)
+            .first(whereFCPElement: .project)
+        
+        let projectName = parentProject?.name
+        
+        // the timeline always needs a name, whether it's a project or a clip.
+        // we prefer the project name if a project exists.
+        // this should also not be an empty string.
+        let timelineName = projectName
+            ?? timeline.timelineName
+            ?? defaultTimelineName
+        
+        // extract from origin element
+        guard let timelineStartTime = timeline.timelineStartAsTimecode() else {
+            logger.error(
+                "Could not determine timeline start timecode."
+            )
+            return nil
+        }
+        
+        return TimelineContext(
+            library: library,
+            projectName: projectName,
+            timeline: timeline,
+            timelineName: timelineName,
+            timelineStartTime: timelineStartTime
+        )
+    }
+    
     public func extractMarkers(
-        preloadedProjects projects: [FinalCutPro.FCPXML.Project]? = nil
+        context: TimelineContext
     ) async -> [Marker] {
         progress.completedUnitCount = 0
         progress.totalUnitCount = 1
@@ -97,37 +152,23 @@ class FCPXMLMarkerExtractor: NSObject, ProgressReporting {
         defer { progress.completedUnitCount = 1 }
         
         var fcpxmlMarkers: [Marker] = []
-
-        let parsedFCPXML = FinalCutPro.FCPXML(fileContent: fcpxmlDoc)
         
-        let library = parsedFCPXML.root.library
+        if markersSource.includesMarkers {
+            fcpxmlMarkers += await markers(
+                in: context.timeline,
+                library: context.library,
+                timelineName: context.timelineName,
+                timelineStartTime: context.timelineStartTime
+            )
+        }
         
-        let projects = projects ?? FinalCutPro.FCPXML(fileContent: fcpxmlDoc)
-            .allProjects()
-        
-        for project in projects {
-            guard let projectStartTime = project.startTimecode() else {
-                logger.error(
-                    "Could not determine start time for project \((project.name ?? "").quoted)."
-                )
-                return []
-            }
-            
-            if markersSource.includesMarkers {
-                fcpxmlMarkers += await markers(
-                    in: project,
-                    library: library,
-                    projectStartTime: projectStartTime
-                )
-            }
-            
-            if markersSource.includesCaptions {
-                fcpxmlMarkers += await captions(
-                    in: project,
-                    library: library,
-                    projectStartTime: projectStartTime
-                )
-            }
+        if markersSource.includesCaptions {
+            fcpxmlMarkers += await captions(
+                in: context.timeline,
+                library: context.library,
+                timelineName: context.timelineName,
+                timelineStartTime: context.timelineStartTime
+            )
         }
         
         // remove markers with excluded roles
@@ -139,13 +180,14 @@ class FCPXMLMarkerExtractor: NSObject, ProgressReporting {
     }
     
     // MARK: - Private Methods
-
+    
     private func markers(
-        in project: FinalCutPro.FCPXML.Project,
+        in timeline: FinalCutPro.FCPXML.AnyTimeline,
         library: FinalCutPro.FCPXML.Library?,
-        projectStartTime: Timecode
+        timelineName: String,
+        timelineStartTime: Timecode
     ) async -> [Marker] {
-        let extractedMarkers = await project.extract(
+        let extractedMarkers = await timeline.extract(
             preset: .markers,
             scope: MarkersExtractor.extractionScope(includeDisabled: includeDisabled)
         )
@@ -154,17 +196,19 @@ class FCPXMLMarkerExtractor: NSObject, ProgressReporting {
             convertMarker(
                 $0,
                 parentLibrary: library,
-                projectStartTime: projectStartTime
+                timelineName: timelineName,
+                timelineStartTime: timelineStartTime
             )
         }
     }
     
     private func captions(
-        in project: FinalCutPro.FCPXML.Project,
+        in timeline: FinalCutPro.FCPXML.AnyTimeline,
         library: FinalCutPro.FCPXML.Library?,
-        projectStartTime: Timecode
+        timelineName: String,
+        timelineStartTime: Timecode
     ) async -> [Marker] {
-        let extractedCaptions = await project.extract(
+        let extractedCaptions = await timeline.extract(
             preset: .captions,
             scope: MarkersExtractor.extractionScope(includeDisabled: includeDisabled)
         )
@@ -173,7 +217,8 @@ class FCPXMLMarkerExtractor: NSObject, ProgressReporting {
             convertCaption(
                 $0,
                 parentLibrary: library,
-                projectStartTime: projectStartTime
+                timelineName: timelineName,
+                timelineStartTime: timelineStartTime
             )
         }
     }
@@ -181,15 +226,17 @@ class FCPXMLMarkerExtractor: NSObject, ProgressReporting {
     private func convertMarker(
         _ extractedMarker: FinalCutPro.FCPXML.ExtractedMarker,
         parentLibrary: FinalCutPro.FCPXML.Library?,
-        projectStartTime: Timecode
+        timelineName: String,
+        timelineStartTime: Timecode
     ) -> Marker? {
         let roles = getClipRoles(extractedMarker)
         
         guard let position = extractedMarker.value(forContext: .absoluteStartAsTimecode()),
               let parentInfo = parentInfo(
                   from: extractedMarker,
-                  parentLibrary: parentLibrary,
-                  projectStartTime: projectStartTime
+                  parentLibrary: parentLibrary, 
+                  timelineName: timelineName,
+                  timelineStartTime: timelineStartTime
               )
         else {
             logger.error("Error converting marker: \(extractedMarker.name.quoted).")
@@ -212,7 +259,8 @@ class FCPXMLMarkerExtractor: NSObject, ProgressReporting {
     private func convertCaption(
         _ extractedCaption: FinalCutPro.FCPXML.ExtractedCaption,
         parentLibrary: FinalCutPro.FCPXML.Library?,
-        projectStartTime: Timecode
+        timelineName: String,
+        timelineStartTime: Timecode
     ) -> Marker? {
         let roles = getClipRoles(extractedCaption)
         
@@ -221,8 +269,9 @@ class FCPXMLMarkerExtractor: NSObject, ProgressReporting {
         guard let position = extractedCaption.timecode(),
               let parentInfo = parentInfo(
                   from: extractedCaption,
-                  parentLibrary: parentLibrary,
-                  projectStartTime: projectStartTime
+                  parentLibrary: parentLibrary, 
+                  timelineName: timelineName,
+                  timelineStartTime: timelineStartTime
               )
         else {
             logger.error("Error converting caption: \(name.quoted).")
@@ -245,7 +294,8 @@ class FCPXMLMarkerExtractor: NSObject, ProgressReporting {
     private func parentInfo(
         from element: any FCPXMLExtractedModelElement,
         parentLibrary: FinalCutPro.FCPXML.Library?,
-        projectStartTime: Timecode
+        timelineName: String,
+        timelineStartTime: Timecode
     ) -> Marker.ParentInfo? {
         guard let clipInTime = element.value(forContext: .parentAbsoluteStartAsTimecode()),
               // let clipDuration = element.value(forContext: .parentDurationAsTimecode()),
@@ -258,10 +308,11 @@ class FCPXMLMarkerExtractor: NSObject, ProgressReporting {
             clipInTime: clipInTime,
             clipOutTime: clipOutTime,
             clipKeywords: element.value(forContext: .keywordsFlat(constrainToKeywordRanges: true)),
+            libraryName: parentLibrary?.name ?? "",
             eventName: element.value(forContext: .ancestorEventName) ?? "",
-            projectName: element.value(forContext: .ancestorProjectName) ?? "",
-            projectStartTime: projectStartTime,
-            libraryName: parentLibrary?.name ?? ""
+            projectName: element.value(forContext: .ancestorProjectName),
+            timelineName: timelineName,
+            timelineStartTime: timelineStartTime
         )
     }