+++ CHANGELOG.md
@@ -2,9 +2,17 @@
 
 ## 0.3.7
 
+### New Features
+
+- Extraction now supports FCPXML containing a standalone clip without a project (#99)
+
 ### Bug Fixes
 
 - Output folder name timestamp now formats as 24-hour time (#100)
+
+### Changes
+
+- Renamed `projectTimecode` naming mode to `timelineNameAndTimecode` (#99)
 
 ## [0.3.6](https://github.com/TheAcharya/MarkersExtractor/releases/tag/0.3.6) (2024-05-05)